pyproject.toml
[tool.poetry]
name = "aluminium-prediction-api"
#packages = [{include = "src"}]
version = "0.1.0"
description = ""
authors = ["KKobuszewski <konrad.kobuszewski93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^1.10.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"
#pandas = "^2.1.1"
#scikit-learn = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
